requirements.txt
PlexAPI==4.15.7
deemix==3.6.6
deezer-py==1.3.7
PyYAML~=6.0.1
